+++ changelog.md
@@ -11,6 +11,7 @@ Codion Change Log
 ### is.codion.common.db
 - AbstractConnectionPoolWrapper.closeStatisticsCollection() added.
 - SLF4J dependency added, DefaultDatabaseConnection logging improved.
+- ConnectionPoolWrapper.DEFAULT_CHECK_OUT_TIMEOUT added.
 ### is.codion.dbms.h2
 - H2Database, security validation for the H2 database script injection.
 ### is.codion.swing.common.ui